Added the new sparse grid. Compiles but needs to be checked.

+++ src/LBgrid.cpp
@@ -4,7 +4,7 @@ Grid::Grid(const int maxNodeNo, const int stride) :maxNodeNo_(maxNodeNo), stride
 {
     // Needs to add one to include a node with label = maxNodeNo_
     neigList_ = new int [(maxNodeNo_ + 1) * stride_];
-    pos_ = new double [(maxNodeNo_ + 1) * 2];
+    pos_ = new int [(maxNodeNo_ + 1) * 2];
 }
 
 Grid::~Grid()
@@ -17,7 +17,7 @@ void Grid::addNeigNode(const int qDirection, const int nodeNo, const int nodeNei
     neigList_[nodeNo * stride_ + qDirection] = nodeNeigNo;
 }
 
-void Grid::addNodePos(const double x, const double y, const int nodeNo)
+void Grid::addNodePos(const int x, const int y, const int nodeNo)
 {
     pos_[2*nodeNo] = x;
     pos_[2*nodeNo + 1] = y;
